extension to biweekly recurrence module
