+++ spec/models/recurrence_spec.rb
@@ -15,90 +15,90 @@ describe Recurrence do
       context 'month created' do
         it 'creates item' do
           Timecop.freeze(Date.new(2016,9,1)) do
-            expect{Recurrence.new(recurring_income).compute}.to change(Income, :count).by(1)
+            expect{Recurring.new(recurring_income).compute}.to change(Income, :count).by(1)
           end
         end
         context 'a day late' do
           it 'creates item' do
             Timecop.freeze(Date.new(2016,9,2)) do
-              expect{Recurrence.new(recurring_income).compute}.to change(Income, :count).by(1)
+              expect{Recurring.new(recurring_income).compute}.to change(Income, :count).by(1)
             end
             Timecop.freeze(Date.new(2016,9,2)) do
-              expect{Recurrence.new(recurring_income).compute}.to change(Income, :count).by(0)
+              expect{Recurring.new(recurring_income).compute}.to change(Income, :count).by(0)
             end
           end
         end
         context '14 days later' do
           it 'creates second item' do
             Timecop.freeze(Date.new(2016,9,1)) do
-              expect{Recurrence.new(recurring_income).compute}.to change(Income, :count).by(1)
+              expect{Recurring.new(recurring_income).compute}.to change(Income, :count).by(1)
             end
 
             Timecop.freeze(Date.new(2016,9,1) + 14) do
-              expect{Recurrence.new(recurring_income).compute}.to change(Income, :count).by(1)
+              expect{Recurring.new(recurring_income).compute}.to change(Income, :count).by(1)
             end
           end
         end
         context 'days in between first time' do
           before do
             Timecop.freeze(Date.new(2016,9,1)) do
-              expect{Recurrence.new(recurring_income).compute}.to change(Income, :count).by(1)
+              expect{Recurring.new(recurring_income).compute}.to change(Income, :count).by(1)
             end
           end
           it 'does not create an item' do
             Timecop.freeze(Date.new(2016,9,1)) do
-              expect{Recurrence.new(recurring_income).compute}.to change(Income, :count).by(0)
+              expect{Recurring.new(recurring_income).compute}.to change(Income, :count).by(0)
             end
             Timecop.freeze(Date.new(2016,9,2)) do
-              expect{Recurrence.new(recurring_income).compute}.to change(Income, :count).by(0)
+              expect{Recurring.new(recurring_income).compute}.to change(Income, :count).by(0)
             end
             Timecop.freeze(Date.new(2016,9,3)) do
-              expect{Recurrence.new(recurring_income).compute}.to change(Income, :count).by(0)
+              expect{Recurring.new(recurring_income).compute}.to change(Income, :count).by(0)
             end
             Timecop.freeze(Date.new(2016,9,4)) do
-              expect{Recurrence.new(recurring_income).compute}.to change(Income, :count).by(0)
+              expect{Recurring.new(recurring_income).compute}.to change(Income, :count).by(0)
             end
             Timecop.freeze(Date.new(2016,9,10)) do
-              expect{Recurrence.new(recurring_income).compute}.to change(Income, :count).by(0)
+              expect{Recurring.new(recurring_income).compute}.to change(Income, :count).by(0)
             end
             Timecop.freeze(Date.new(2016,9,13)) do
-              expect{Recurrence.new(recurring_income).compute}.to change(Income, :count).by(0)
+              expect{Recurring.new(recurring_income).compute}.to change(Income, :count).by(0)
             end
             Timecop.freeze(Date.new(2016,9,14)) do
-              expect{Recurrence.new(recurring_income).compute}.to change(Income, :count).by(0)
+              expect{Recurring.new(recurring_income).compute}.to change(Income, :count).by(0)
             end
           end
         end
         context 'days in between second time' do
           before do
             Timecop.freeze(Date.new(2016,9,1)) do
-              expect{Recurrence.new(recurring_income).compute}.to change(Income, :count).by(1)
+              expect{Recurring.new(recurring_income).compute}.to change(Income, :count).by(1)
             end
             Timecop.freeze(Date.new(2016,9,1) + 14) do
-              expect{Recurrence.new(recurring_income).compute}.to change(Income, :count).by(1)
+              expect{Recurring.new(recurring_income).compute}.to change(Income, :count).by(1)
             end
           end
           it 'does not create an item' do
             Timecop.freeze(Date.new(2016,9,15)) do
-              expect{Recurrence.new(recurring_income).compute}.to change(Income, :count).by(0)
+              expect{Recurring.new(recurring_income).compute}.to change(Income, :count).by(0)
             end
             Timecop.freeze(Date.new(2016,9,16)) do
-              expect{Recurrence.new(recurring_income).compute}.to change(Income, :count).by(0)
+              expect{Recurring.new(recurring_income).compute}.to change(Income, :count).by(0)
             end
             Timecop.freeze(Date.new(2016,9,17)) do
-              expect{Recurrence.new(recurring_income).compute}.to change(Income, :count).by(0)
+              expect{Recurring.new(recurring_income).compute}.to change(Income, :count).by(0)
             end
             Timecop.freeze(Date.new(2016,9,18)) do
-              expect{Recurrence.new(recurring_income).compute}.to change(Income, :count).by(0)
+              expect{Recurring.new(recurring_income).compute}.to change(Income, :count).by(0)
             end
             Timecop.freeze(Date.new(2016,9,25)) do
-              expect{Recurrence.new(recurring_income).compute}.to change(Income, :count).by(0)
+              expect{Recurring.new(recurring_income).compute}.to change(Income, :count).by(0)
             end
             Timecop.freeze(Date.new(2016,9,29)) do
-              expect{Recurrence.new(recurring_income).compute}.to change(Income, :count).by(0)
+              expect{Recurring.new(recurring_income).compute}.to change(Income, :count).by(0)
             end
             Timecop.freeze(Date.new(2016,9,30)) do
-              expect{Recurrence.new(recurring_income).compute}.to change(Income, :count).by(0)
+              expect{Recurring.new(recurring_income).compute}.to change(Income, :count).by(0)
             end
           end
         end
@@ -107,70 +107,70 @@ describe Recurrence do
         context 'days in between first time' do
           before do
             Timecop.freeze(Date.new(2016,9,1)) do
-              expect{Recurrence.new(recurring_income).compute}.to change(Income, :count).by(1)
+              expect{Recurring.new(recurring_income).compute}.to change(Income, :count).by(1)
             end
             Timecop.freeze(Date.new(2016,9,15)) do
-              expect{Recurrence.new(recurring_income).compute}.to change(Income, :count).by(1)
+              expect{Recurring.new(recurring_income).compute}.to change(Income, :count).by(1)
             end
 
             Timecop.freeze(Date.new(2016,10,1)) do
-              expect{Recurrence.new(recurring_income).compute}.to change(Income, :count).by(1)
+              expect{Recurring.new(recurring_income).compute}.to change(Income, :count).by(1)
             end
           end
           it 'does not create an item' do
             Timecop.freeze(Date.new(2016,10,1)) do
-              expect{Recurrence.new(recurring_income).compute}.to change(Income, :count).by(0)
+              expect{Recurring.new(recurring_income).compute}.to change(Income, :count).by(0)
             end
             Timecop.freeze(Date.new(2016,10,2)) do
-              expect{Recurrence.new(recurring_income).compute}.to change(Income, :count).by(0)
+              expect{Recurring.new(recurring_income).compute}.to change(Income, :count).by(0)
             end
             Timecop.freeze(Date.new(2016,10,3)) do
-              expect{Recurrence.new(recurring_income).compute}.to change(Income, :count).by(0)
+              expect{Recurring.new(recurring_income).compute}.to change(Income, :count).by(0)
             end
             Timecop.freeze(Date.new(2016,10,4)) do
-              expect{Recurrence.new(recurring_income).compute}.to change(Income, :count).by(0)
+              expect{Recurring.new(recurring_income).compute}.to change(Income, :count).by(0)
             end
             Timecop.freeze(Date.new(2016,10,10)) do
-              expect{Recurrence.new(recurring_income).compute}.to change(Income, :count).by(0)
+              expect{Recurring.new(recurring_income).compute}.to change(Income, :count).by(0)
             end
             Timecop.freeze(Date.new(2016,10,13)) do
-              expect{Recurrence.new(recurring_income).compute}.to change(Income, :count).by(0)
+              expect{Recurring.new(recurring_income).compute}.to change(Income, :count).by(0)
             end
             Timecop.freeze(Date.new(2016,10,14)) do
-              expect{Recurrence.new(recurring_income).compute}.to change(Income, :count).by(0)
+              expect{Recurring.new(recurring_income).compute}.to change(Income, :count).by(0)
             end
           end
         end
         context 'days in between second time' do
           before do
             Timecop.freeze(Date.new(2016,10,1)) do
-              expect{Recurrence.new(recurring_income).compute}.to change(Income, :count).by(1)
+              expect{Recurring.new(recurring_income).compute}.to change(Income, :count).by(1)
             end
             Timecop.freeze(Date.new(2016,10,1) + 14) do
-              expect{Recurrence.new(recurring_income).compute}.to change(Income, :count).by(1)
+              expect{Recurring.new(recurring_income).compute}.to change(Income, :count).by(1)
             end
           end
           it 'does not create an item' do
             Timecop.freeze(Date.new(2016,10,15)) do
-              expect{Recurrence.new(recurring_income).compute}.to change(Income, :count).by(0)
+              expect{Recurring.new(recurring_income).compute}.to change(Income, :count).by(0)
             end
             Timecop.freeze(Date.new(2016,10,16)) do
-              expect{Recurrence.new(recurring_income).compute}.to change(Income, :count).by(0)
+              expect{Recurring.new(recurring_income).compute}.to change(Income, :count).by(0)
             end
             Timecop.freeze(Date.new(2016,10,17)) do
-              expect{Recurrence.new(recurring_income).compute}.to change(Income, :count).by(0)
+              expect{Recurring.new(recurring_income).compute}.to change(Income, :count).by(0)
             end
             Timecop.freeze(Date.new(2016,10,18)) do
-              expect{Recurrence.new(recurring_income).compute}.to change(Income, :count).by(0)
+              expect{Recurring.new(recurring_income).compute}.to change(Income, :count).by(0)
             end
             Timecop.freeze(Date.new(2016,10,25)) do
-              expect{Recurrence.new(recurring_income).compute}.to change(Income, :count).by(0)
+              expect{Recurring.new(recurring_income).compute}.to change(Income, :count).by(0)
             end
             Timecop.freeze(Date.new(2016,10,29)) do
-              expect{Recurrence.new(recurring_income).compute}.to change(Income, :count).by(0)
+              expect{Recurring.new(recurring_income).compute}.to change(Income, :count).by(0)
             end
             Timecop.freeze(Date.new(2016,10,30)) do
-              expect{Recurrence.new(recurring_income).compute}.to change(Income, :count).by(0)
+              expect{Recurring.new(recurring_income).compute}.to change(Income, :count).by(0)
             end
           end
         end
@@ -192,9 +192,9 @@ describe Recurrence do
     context 'when due' do
       context 'a month later' do
         it 'creates a pending expense' do
-          expect{Recurrence.new(recurring_expense).compute}.to change(PendingExpense.unscoped, :count).by(1)
+          expect{Recurring.new(recurring_expense).compute}.to change(PendingExpense.unscoped, :count).by(1)
           Timecop.freeze(Date.today + 1.month) do
-            expect{Recurrence.new(recurring_expense).compute}.to change(PendingExpense.unscoped, :count).by(1)
+            expect{Recurring.new(recurring_expense).compute}.to change(PendingExpense.unscoped, :count).by(1)
           end
           expect(PendingExpense.unscoped.count).to be 2
         end
@@ -202,24 +202,24 @@ describe Recurrence do
       context 'when duplicate' do
         context 'on same day' do
           it 'does not create' do
-            expect{Recurrence.new(recurring_expense).compute}.to change(PendingExpense.unscoped, :count).by(1)
-            expect{Recurrence.new(recurring_expense).compute}.to change(PendingExpense.unscoped, :count).by(0)
+            expect{Recurring.new(recurring_expense).compute}.to change(PendingExpense.unscoped, :count).by(1)
+            expect{Recurring.new(recurring_expense).compute}.to change(PendingExpense.unscoped, :count).by(0)
             expect(PendingExpense.unscoped.count).to be 1
           end
         end
         context 'in the following month after logging the first time the item was due' do
           before do
-            expect{Recurrence.new(recurring_expense).compute}.to change(PendingExpense.unscoped, :count).by(1)
+            expect{Recurring.new(recurring_expense).compute}.to change(PendingExpense.unscoped, :count).by(1)
             Timecop.freeze(Date.today + 1.month) do
-              expect{Recurrence.new(recurring_expense).compute}.to change(PendingExpense.unscoped, :count).by(1)
+              expect{Recurring.new(recurring_expense).compute}.to change(PendingExpense.unscoped, :count).by(1)
             end
           end
           it 'does not create again on the following day' do
             Timecop.freeze(Date.today + 1.month + 1.day) do
-              expect{Recurrence.new(recurring_expense).compute}.to change(PendingExpense.unscoped, :count).by(0)
+              expect{Recurring.new(recurring_expense).compute}.to change(PendingExpense.unscoped, :count).by(0)
             end
             Timecop.freeze(Date.today + 1.month + 2.days) do
-              expect{Recurrence.new(recurring_expense).compute}.to change(PendingExpense.unscoped, :count).by(0)
+              expect{Recurring.new(recurring_expense).compute}.to change(PendingExpense.unscoped, :count).by(0)
             end
             expect(PendingExpense.unscoped.count).to be 2
           end
@@ -232,19 +232,19 @@ describe Recurrence do
         # Timecop.freeze (yyyy, mm, dd)
         recurring_expense.update(created_at: Time.new(2016,9,10))
         Timecop.freeze(2016,10,1) do
-          expect{Recurrence.new(recurring_expense).compute}.to change(PendingExpense.unscoped, :count).by(0)
+          expect{Recurring.new(recurring_expense).compute}.to change(PendingExpense.unscoped, :count).by(0)
         end
         Timecop.freeze(2016,10,5) do
-          expect{Recurrence.new(recurring_expense).compute}.to change(PendingExpense.unscoped, :count).by(0)
+          expect{Recurring.new(recurring_expense).compute}.to change(PendingExpense.unscoped, :count).by(0)
         end
         Timecop.freeze(2016,10,10) do
-          expect{Recurrence.new(recurring_expense).compute}.to change(PendingExpense.unscoped, :count).by(1)
+          expect{Recurring.new(recurring_expense).compute}.to change(PendingExpense.unscoped, :count).by(1)
         end
         Timecop.freeze(2016,10,11) do
-          expect{Recurrence.new(recurring_expense).compute}.to change(PendingExpense.unscoped, :count).by(0)
+          expect{Recurring.new(recurring_expense).compute}.to change(PendingExpense.unscoped, :count).by(0)
         end
         Timecop.freeze(2016,10,25) do
-          expect{Recurrence.new(recurring_expense).compute}.to change(PendingExpense.unscoped, :count).by(0)
+          expect{Recurring.new(recurring_expense).compute}.to change(PendingExpense.unscoped, :count).by(0)
         end
       end
     end
@@ -265,19 +265,19 @@ describe Recurrence do
     context 'when not due' do
       before do
         Timecop.freeze(time) do
-          expect{Recurrence.new(recurring_expense).compute}.to change(PendingExpense.unscoped, :count).by(1)
+          expect{Recurring.new(recurring_expense).compute}.to change(PendingExpense.unscoped, :count).by(1)
         end
       end
 
       it 'does not log in same year' do
         Timecop.freeze(time.tomorrow) do
-          expect{Recurrence.new(recurring_expense).compute}.to change(PendingExpense.unscoped, :count).by(0)
+          expect{Recurring.new(recurring_expense).compute}.to change(PendingExpense.unscoped, :count).by(0)
         end
       end
 
       it 'does not log in next year before month due' do
         Timecop.freeze(time.next_year.beginning_of_year) do
-          expect{Recurrence.new(recurring_expense).compute}.to change(PendingExpense.unscoped, :count).by(0)
+          expect{Recurring.new(recurring_expense).compute}.to change(PendingExpense.unscoped, :count).by(0)
         end
       end
     end
@@ -285,14 +285,14 @@ describe Recurrence do
     context 'when due' do
       before do
         Timecop.freeze(time) do
-          expect{Recurrence.new(recurring_expense).compute}.to change(PendingExpense.unscoped, :count).by(1)
+          expect{Recurring.new(recurring_expense).compute}.to change(PendingExpense.unscoped, :count).by(1)
         end
       end
 
 
       it 'logs' do
         Timecop.freeze(time.next_year) do
-          expect{Recurrence.new(recurring_expense).compute}.to change(PendingExpense.unscoped, :count).by(1)
+          expect{Recurring.new(recurring_expense).compute}.to change(PendingExpense.unscoped, :count).by(1)
         end
       end
     end
